Fix import sort order
When reformatting imports, rustfmt takes into consideration the casing
of the item imported and treats all uppercase items with preference. We
do not want such a special treatment. Items should be sorted based
solely on the letter and not its casing.
This change adjusts rustfmt to change sorting of imports to this simpler
scheme.