Format match arms in vertical style
By default rustfmt formats match arms in a way to prefer horizontal
layout over vertical one. E.g.,
  > match key {
  >   Key::Esc | Key::Return => {
  >     ...
  >   },

Most of the time, however, a vertical layout is preferred as it allows
for easier addition/removal of lines and generally (although
subjectively) provides for a quicker overview of the alternatives. For
example:
  > match key {
  >   Key::Esc |
  >   Key::Return => {
  >     self.in_out = InOut::Clear;

This change adjusts rustfmt to format match arms in such a vertical
style.