End attributes with a trailing comma
With changes leading up to 1.0.0, rustfmt removes the trailing comma
from attribute lists such as:
  > #![deny(
  >   unused_results,
  >   warnings,
  > )]

This behavior is undesired, as it causes unnecessary churn in the diff
when adding a new element to the end of the attribute list.
With this patch we bring back the original behavior of keeping (or
adding) the trailing comma.